WebApp/requirements.txt
Flask==1.1.1
Flask-Cors==3.0.8
psycopg2==2.8.6
pycryptodome==3.10.1